Move identity's user lookup into the queries package
`GetUserByIdUseCase` is a pure read: one `find_by_id`, one NotFound, no
mutation. Per ADR-0001 it belongs in `queries/`, not in `use_cases/`.

It stays on the halfway option — same repository, same `User` entity —
rather than growing a query port and a view DTO. Two reasons:

- `user_repository.find_by_id` is shared with `UpdateUserUseCase` and
  `RefreshAccessTokenUseCase`, so a port would remove nothing from the
  write side; the guide's "ask what the port would remove" test says
  stay minimal.
- Its only consumer is `get_current_user`, and that `User` reaches every
  router in the app, which passes its id into command use cases. A read
  DTO here would be a Rule 2 violation by construction.
