+++ backend/src/application/identity/queries/__init__.py
@@ -1,0 +1,1 @@
+"""Identity read use cases (query side)."""
