+++ backend/src/application/identity/use_cases/__init__.py
@@ -1,12 +1,10 @@
 from .authentication.authenticate_user_use_case import AuthenticateUserUseCase
-from .authentication.get_user_by_id_use_case import GetUserByIdUseCase
 from .authentication.refresh_access_token_use_case import RefreshAccessTokenUseCase
 from .register_user_use_case import RegisterUserUseCase
 from .update_user_use_case import UpdateUserUseCase
 
 __all__ = [
     "AuthenticateUserUseCase",
-    "GetUserByIdUseCase",
     "RefreshAccessTokenUseCase",
     "RegisterUserUseCase",
     "UpdateUserUseCase",
